feat: reestructurar middleware para manejar rutas de autenticación y autorización
