--- frontend/middleware.ts
@@ -1,31 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-import jwt from "jsonwebtoken";
-
-const JWT_SECRET = process.env.JWT_SECRET;
-
-if (!JWT_SECRET) {
-  throw new Error("JWT_SECRET no está definido en las variables de entorno");
-}
-
-export default function middleware(req: NextRequest) {
-  const token = req.cookies.get("authToken")?.value;
-
-  if (!token) {
-    return NextResponse.redirect(new URL("/", req.url));
-  }
-  try {
-    const decoded = jwt.verify(token, JWT_SECRET) as unknown as {
-      userId: string;
-      role: string;
-    };
-    if (decoded.role !== "admin" && req.nextUrl.pathname.startsWith("/admin")) {
-      return NextResponse.redirect(new URL("/", req.url));
-    }
-  } catch (error) {
-    return NextResponse.redirect(new URL("/", req.url));
-  }
-}
-/* if (!token && protectedRoutes.includes(req.nextUrl.pathname)) {
-    const absoluteURL = new URL("/", req.nextUrl.origin);
-    return NextResponse.redirect(absoluteURL.toString());
-  } */
