+++ frontend/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.3.0",
     "@types/jsonwebtoken": "^9.0.7",
+    "dotenv": "^16.4.7",
     "html-react-parser": "^5.1.19",
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^4.0.0",
