+++ frontend/src/app/admin/libros/libro/crear/page.tsx
@@ -1,0 +1,7 @@
+export default function NewBookPage() {
+  return (
+    <div>
+      <h1>Hello Page</h1>
+    </div>
+  );
+}
